+++ draw_2d/vector/font/src/horizontal_metrics.rs
@@ -1,6 +1,6 @@
 /// The horizontal metrics for a glyph
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct HorizontalMetrics {
-    pub advance_width: f32,
-    pub left_side_bearing: f32,
+    pub advance_width: f64,
+    pub left_side_bearing: f64,
 }
